Iteration 4 Move tests to the rss_reader folder; Add prints when html and pdf documents were created
