+++ final_task/rss_reader/tests.py
@@ -1,9 +1,5 @@
 import os
-import sys
 import unittest
-
-out_dir = os.path.abspath(os.path.dirname(os.path.dirname(__file__)))
-sys.path.append(out_dir)
 from Handler import Handler
 from Entry import Entry
 from RSSReaderException import RSSReaderException
@@ -163,7 +159,7 @@ class TestHandler(unittest.TestCase):
         open_mock.assert_called_with(f"{os.path.abspath(os.path.dirname(__file__))}\\RSS_News.html", "w", encoding='utf-8')
         open_mock.return_value.write.assert_called_once()
         open_mock.return_value.write.assert_called_once_with('<html>\n  <head>\n    <meta charset="utf-8">\n  </head>\n  <body>\n    <div>\n      <h1>title</h1>\n      <p>\n        <b>Feed: </b>\n        <a>feed</a>\n      </p>\n      <p>\n        <b>Date: </b>\n        <a>Wed, 20 Nov 2019</a>\n      </p>\n      <img src="file:///F:\\Introduction to Python\\Final Task\\FinalTaskRssParser\\final_task\\rss_reader/images/title0.jpg"><br><br>\n      <p>some_text<br><br></p>\n    </div>\n  </body>\n</html>')
-                                                             
+
     def test_write_entries_to_html_without_img(self):
         open_mock = mock_open()
         entry = Entry(feed="feed", title="title", date="Wed, 20 Nov 2019", article_link="link",
@@ -195,9 +191,98 @@ class TestHandler(unittest.TestCase):
         self.handler.print_to_json = MagicMock(return_value=True)
         self.handler.get_entry_from_dict = MagicMock(return_value=True)
         self.handler.print_entry = MagicMock(return_value=True)
-        self.handler.option_date("20191111", True, "path_html", "path_pdf")
-        self.handler.option_date("20191111", False)
-        # self.assertEqual(True, self.handler.option_date("20191111", True, "path_html", "path_pdf"))
+        # self.handler.option_date("20191111", False)
+
+
+class TestEntry(unittest.TestCase):
+    def test__init__(self):
+        entry = Entry()
+        self.assertIsInstance(entry, Entry)
+        self.assertNotIsInstance("entry", Entry)
+
+    def setUp(self):
+        self.entry = Entry(feed="feed", title="title", date="Wed, 20 Nov 2019", article_link="link",
+                           summary="some_text", links=("article_link", "img_link"))
+
+    def test_parse_html(self):
+        self.assertEqual(self.entry.parse_html
+            (
+            "Graham now says Trump&#39;s Ukraine policy was too &#39;incoherent&#39; for quid pro quo"
+        ), "Graham now says Trump's Ukraine policy was too 'incoherent' for quid pro quo"
+        )
+        self.assertEqual(self.entry.parse_html(
+            '<p><a href="https://news.yahoo.com/graham-trump-ukraine-incoherent-quid-pro-quo-192210175.html">'
+            '<img src="http://l2.yimg.com/uu/api/res/1.2/aWhGys7_IW5qIjKaiJpPfg--/YXBwaWQ9eXRhY2h5b247aD04Njt'
+            '3PTEzMDs-/https://media-mbst-pub-ue1.s3.amazonaws.com/creatr-uploaded-images/2019-11/5527ffe0-00'
+            'ca-11ea-9f7d-d1e736c1315d" width="130" height="86" alt="Graham now says Trump&#39;s Ukraine poli'
+            'cy was too &#39;incoherent&#39; for quid pro quo" align="left" title="Graham now says Trump&#39;'
+            's Ukraine policy was too &#39;incoherent&#39; for quid pro quo" border="0" ></a>A day after sayi'
+            'ng he wouldn’t bother to read the testimony, Sen. Lindsey Graham now says he did read it, and hi'
+            's conclusion is that the Trump administration’s Ukraine policy was too &quot;incoherent&quot; fo'
+            'r it to have orchestrated the quid pro quo at the heart of the impeachment inquiry.<p><br clear="all">'
+        ),
+            "[image 2: Graham now says Trump's Ukraine policy was too 'incoherent' for quid pro quo][2] A day after "
+            "saying he wouldn’t bother to read the testimony, Sen. Lindsey Graham now says he did read it, and his "
+            "conclusion is that the Trump administration’s Ukraine policy was too \"incoherent\" for it to have"
+            " orchestrated the quid pro quo at the heart of the impeachment inquiry."
+        )
+        self.assertNotEqual(self.entry.parse_html(
+            "Graham now says Trump&#39;s Ukraine policy was too &#39;incoherent&#39; for quid pro quo"
+        ),
+            "Ukraine policy was too 'incoherent' for quid pro quo"
+        )
+        self.assertNotEqual(self.entry.parse_html(
+            '<p><a href="https://news.yahoo.com/graham-trump-ukraine-incoherent-quid-pro-quo-192210175.html">'
+            '<img src="http://l2.yimg.com/uu/api/res/1.2/aWhGys7_IW5qIjKaiJpPfg--/YXBwaWQ9eXRhY2h5b247aD04Njt'
+            '3PTEzMDs-/https://media-mbst-pub-ue1.s3.amazonaws.com/creatr-uploaded-images/2019-11/5527ffe0-00'
+            'ca-11ea-9f7d-d1e736c1315d" width="130" height="86" alt="Graham now says Trump&#39;s Ukraine poli'
+            'cy was too &#39;incoherent&#39; for quid pro quo" align="left" title="Graham now says Trump&#39;'
+            's Ukraine policy was too &#39;incoherent&#39; for quid pro quo" border="0" ></a>A day after sayi'
+            'ng he wouldn’t bother to read the testimony, Sen. Lindsey Graham now says he did read it, and hi'
+            's conclusion is that the Trump administration’s Ukraine policy was too &quot;incoherent&quot; fo'
+            'r it to have orchestrated the quid pro quo at the heart of the impeachment inquiry.<p><br clear="all">'
+        ),
+            "[image 5: Graham now says Trump's Ukraine policy was too 'incoherent' for quid pro quo][5] A day after "
+            "saying he wouldn’t bother to read the testimony, Sen. Lindsey Graham now says he did read it, and his "
+            "conclusion is that the Trump administration’s Ukraine policy was too \"incoherent\" for it to have"
+            " orchestrated the quid pro quo at the heart of the impeachment inquiry."
+        )
+
+    def test_print_feed(self):
+        expected_feed = 'Feed: feed\n\n'
+        with patch('sys.stdout', new=StringIO()) as fake_out:
+            self.entry.print_feed()
+            self.assertEqual(expected_feed, fake_out.getvalue())
+
+    def test_print_title(self):
+        expected_title = 'Title: title\n'
+        with patch('sys.stdout', new=StringIO()) as fake_out:
+            self.entry.print_title()
+            self.assertEqual(expected_title, fake_out.getvalue())
+
+    def test_print_date(self):
+        expected_date = 'Date: Wed, 20 Nov 2019\n'
+        with patch('sys.stdout', new=StringIO()) as fake_out:
+            self.entry.print_date()
+            self.assertEqual(expected_date, fake_out.getvalue())
+
+    def test_print_link(self):
+        expected_link = 'Link: link\n'
+        with patch('sys.stdout', new=StringIO()) as fake_out:
+            self.entry.print_link()
+            self.assertEqual(expected_link, fake_out.getvalue())
+
+    def test_print_summary(self):
+        expected_summary = '\nsome_text\n\n'
+        with patch('sys.stdout', new=StringIO()) as fake_out:
+            self.entry.print_summary()
+            self.assertEqual(expected_summary, fake_out.getvalue())
+
+    def test_print_links(self):
+        expected_links = 'Links:\n[0] article_link (link)\n[1] img_link (image)\n\n\n'
+        with patch('sys.stdout', new=StringIO()) as fake_out:
+            self.entry.print_links()
+            self.assertEqual(expected_links ,fake_out.getvalue())
 
 
 if __name__ == '__main__':
